Update README.md
I have updated the music player app to include a new random play mode. To access this mode, simply click on the 'Random' button. This button allows the user to switch between random play mode and normal mode. With this new feature, the user can now enjoy listening to their music in a more unpredictable and exciting way.